Restore v7.0.1 Toolkit launcher and menu
