+++ .github/scripts/test_issue454_preboot_state_order.py
@@ -13,6 +13,7 @@ assert 'delete merged.operationalWindow;' in text
 assert 'delete merged.missionRequirements;' in text
 meta=re.search(r'(?m)^//\s*@version\s+([^\s]+)\s*$',text).group(1)
 runtime=re.search(r"version:\s*'([^']+)',",text).group(1)
-assert meta==runtime=='7.0.0'
+assert meta==runtime
+assert tuple(int(part) for part in meta.split('.')[:3]) >= (7,0,0)
 assert text.rstrip().endswith('})();')
 print('v7 preboot hydration contract passed.')
