--- .github/workflows/apply-issue515-launcher-fix.yml
@@ -1,86 +1,0 @@
-name: Apply Issue 515 Emergency Launcher Restoration
-
-on:
-  pull_request:
-    branches:
-      - main
-    paths:
-      - ".github/issue515/apply_launcher_fix.py"
-      - ".github/issue515/patch_candidate_ledgers.py"
-      - ".github/scripts/test_issue464_launcher_settings_contract.py"
-      - ".github/workflows/apply-issue515-launcher-fix.yml"
-
-permissions:
-  contents: read
-
-concurrency:
-  group: toolkit-issue-515-launcher-restoration
-  cancel-in-progress: false
-
-jobs:
-  restore:
-    name: Restore and validate v7.0.1 launcher
-    if: github.actor == 'Conroy1988' && github.head_ref == 'fix/issue-515-v7-launcher-restoration'
-    runs-on: ubuntu-latest
-    timeout-minutes: 30
-    steps:
-      - name: Check out exact emergency branch
-        uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7
-        with:
-          ref: fix/issue-515-v7-launcher-restoration
-          fetch-depth: 0
-          persist-credentials: false
-
-      - name: Apply exact launcher restoration
-        shell: bash
-        run: |
-          set -euo pipefail
-          python3 .github/issue515/apply_launcher_fix.py
-          python3 .github/issue515/patch_candidate_ledgers.py
-
-      - name: Validate complete v7.0.1 candidate
-        id: validation
-        shell: bash
-        run: |
-          set +e
-          set -o pipefail
-          {
-            bash .github/scripts/run_userscript_preflight.sh --all
-            python3 .github/scripts/test_documentation_consistency.py \
-              --allow-release-candidate \
-              --json-output documentation-drift-report.json \
-              --markdown-output documentation-drift-report.md
-          } 2>&1 | tee issue515-launcher-validation.log
-          STATUS=${PIPESTATUS[0]}
-          echo "exit_code=$STATUS" >> "$GITHUB_OUTPUT"
-          exit 0
-
-      - name: Upload exact launcher validation diagnostic
-        if: always()
-        uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
-        with:
-          name: issue515-launcher-validation-${{ github.sha }}
-          path: issue515-launcher-validation.log
-          if-no-files-found: error
-          retention-days: 14
-
-      - name: Stop on launcher validation failure
-        if: steps.validation.outputs.exit_code != '0'
-        run: exit 1
-
-      - name: Commit exact emergency candidate
-        env:
-          DEVELOPMENT_PR_TOKEN: ${{ secrets.DEVELOPMENT_PR_TOKEN }}
-        shell: bash
-        run: |
-          set -euo pipefail
-          test -n "$DEVELOPMENT_PR_TOKEN"
-          test "$GITHUB_HEAD_REF" = "fix/issue-515-v7-launcher-restoration"
-          git config user.name "Conroy1988"
-          git config user.email "27301455+Conroy1988@users.noreply.github.com"
-          git add -A
-          git diff --cached --quiet && { echo "No launcher restoration changes produced."; exit 1; }
-          git commit -m "Restore v7.0.1 Toolkit launcher and menu"
-          BASIC_AUTH="$(printf 'x-access-token:%s' "$DEVELOPMENT_PR_TOKEN" | base64 -w0)"
-          git -c http.https://github.com/.extraheader="AUTHORIZATION: basic ${BASIC_AUTH}" \
-            push origin HEAD:refs/heads/fix/issue-515-v7-launcher-restoration
